Move `create_config` back to hyfetch.rs
p/s: sorry...
